Refactor stories and story component
